+++ components/GroupGrid/demos/index.js
@@ -1,0 +1,13 @@
+define([
+  './basic.js',
+  './add-default-value.js',
+  './events.js',
+  './field-hidden.js',
+  './field-width.js',
+], function () {
+  return {
+    title: 'GroupGrid',
+    subtitle: '字段组表格',
+    demos: Array.prototype.slice.call(arguments),
+  }
+})